Removed comment and an empty line
It's a small change, but it's necessary! Maybe! I think I just don't want it to bother me later.
I accidentally put my Spigot username instead of my Github name on the comment, so I was gonna fix it, but then I just decided to remove it and add it when I make a commenting pass over the whole codebase later.