Created new class "MetaHandler" to handle metadata
This new class will handle most metadata (read: metadata not related to a specific class) to cut down on unnecessary methods in StatusCheck and make it easier to find methods by having them under an appropriate label. It will also help make it easier to debug StatusCheck since I won't have to be making sure it's not a Metadata problem happening in StatusCheck.